refactor(S01): extract scan_runtime.py, wire PA001 into CLI, extract schema loader
T02: Move scan expansion and validation-loop orchestration from
plugin_validator.py into scan_runtime.py. Functions extracted:
_discover_validatable_paths, _resolve_filter_and_expand_paths,
_load_ignore_patterns, _is_ignored, _compute_summary, run_validation_loop.
Circular imports avoided via callback injection.

PA001: Wire PluginAgentFrontmatterValidator into _get_validators_for_path()
so skilllint check actually flags prohibited plugin agent frontmatter fields.

Schema loader: Move schema loading functions from schemas/__init__.py to
schemas/_loader.py (RUF067 fix — functions in __init__).
